use typedheader
