+++ objectstore-server/Cargo.toml
@@ -15,6 +15,7 @@ anyhow = { workspace = true }
 argh = { workspace = true }
 async-stream = { workspace = true }
 axum = { workspace = true, features = ["multipart"] }
+axum-extra = { workspace = true, features = ["typed-header"] }
 base64 = { workspace = true }
 bytes = { workspace = true }
 ed25519-dalek = { workspace = true, features = ["pem"] }
